requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
opencv-python==4.8.0
mediapipe==0.10.9
numpy==1.23.5
pillow==9.5.0
